+++ HISTORY.md
@@ -1,6 +1,25 @@
 FORTPY: Revision History
 ======
 
+Revision 1.0.7
+------
+
+- Added `timed="true"` attribute support to the `<test>` tag.
+- Fixed bug where an empty folder being cleaned would raise an exception on `remove()`
+- Added a check for executables being tested to make sure they have a `public` modifier or are in the list of public symbols in the parent module.
+- Added debug compile option to unit testing.
+- Fixed automatic allocation errors for pointers.
+- Fixed fortpy.f90 line and value counting routines.
+- Added support for arrays of derived types; still get suggestions for embedded members if operating on an element of an array of derived types.
+- Fixed a bug in `parsers.types` where "endtype" was not recognized.
+- Added `<isense>` section to `config.xml` with an option to uppercase the builtin Fortran names
+  by default with the suggestions.
+
+Revision 1.0.6
+------
+
+- Fixed issue with fortpy/templates directory not being installed.
+
 Revision 1.0.5
 ------